FrontEnd progress
A stable versioon that at the moment:
- has a detailed view for a flight, including rendering the path.
- has an admin detailed view: allows to change the status of the flight (fully implemented)

Decigions made:
all controllers made in api style has been renamed to reflect that. An additional layer is going to be created for serving html rather than Json. As the result the web application would be accesible as both api and website.

(so you can alreeady upload an xml file, register, upload igc file and view it by passing  flight ID found in an upload response or in a list method; by logging in as superadmin (look in the appsettings) you could also change the status)